1. Database fixes:    - fixed GameDataBase to actually save deletions    - added equals() and hashCode() to GameRecord
2. UI Improvements:
   - added a "Confirm Delete" popup window
   - Gallery refreshes automatically after deleting an image.

+++ src/main/java/com/sketchandguess/gui/PictureWindow.java
@@ -9,6 +9,7 @@ import javax.swing.*;
 import java.awt.*;
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
+import javax.swing.JOptionPane;
 
 public class PictureWindow extends JFrame implements PropertyChangeListener {
 
@@ -67,8 +68,17 @@ public class PictureWindow extends JFrame implements PropertyChangeListener {
             GameRecord record = state.getCurrentRecord();
 
             if (record != null) {
-                GalleryWindowController.deleteGame();
-                this.dispose();
+                int result = JOptionPane.showConfirmDialog(
+                        this,
+                        "Are you sure you want to delete this image?",
+                        "Confirm Delete",
+                        JOptionPane.YES_NO_OPTION,
+                        JOptionPane.WARNING_MESSAGE
+                );
+                if (result == JOptionPane.YES_OPTION) {
+                    GalleryWindowController.deleteGame();
+                    this.dispose();
+                }
             }
         });
     }
